+++ prompt.h
@@ -5,12 +5,15 @@
 #include <uchar.h>
 
 struct prompt;
-struct prompt *prompt_init(const char32_t *prompt_text);
+struct prompt *prompt_init(
+    const char32_t *prompt_text, const char32_t *placeholder,
+    const char32_t *text);
 void prompt_destroy(struct prompt *prompt);
 
 bool prompt_insert_chars(struct prompt *prompt, const char *text, size_t len);
 
 const char32_t *prompt_prompt(const struct prompt *prompt);
+const char32_t *prompt_placeholder(const struct prompt *prompt);
 const char32_t *prompt_text(const struct prompt *prompt);
 size_t prompt_cursor(const struct prompt *prompt);